externalize authentication module
